Consolidate FeatureOutputStream classes
